Extract cross-BSD OperatingSystem super-base
The FreeBSD, OpenBSD, DragonFly, and NetBSD OperatingSystem bases each carried
identical copies of the cross-BSD-common methods: the Unix/BSD manufacturer, the
uname-based bitness, queryAllProcesses/queryChild/queryDescendant/getProcess, the
ps -axo pid process count, the ps -axo nlwp thread count, the /etc/rc.d services,
and the memoized system uptime/boot time.

Lift those into a shared BsdOperatingSystem super-base (extends
AbstractOperatingSystem). The four platforms now extend it and provide only what
genuinely differs:

- getProcessListFromPS(int) (abstract hook): FreeBSD/DragonFly stream + VALID_PROCESS
  (DragonFly also drops kernel threads with pid <= 0); OpenBSD/NetBSD for-loop.
- queryBootTime() (abstract hook): FreeBSD/DragonFly read a native sysctl struct
  (kept abstract through to the JNA/FFM subclasses); OpenBSD/NetBSD text-parse it.
- getThreadCount(): default ps -axo nlwp sum in the super-base; OpenBSD overrides
  with ps -axHo tid.
- getCurrentThread() and the native version/file-system/network/sysctl methods stay
  per-platform.

NetBSD has no native-access split, so its single class extends the super-base
directly. The JNA/FFM subclasses are unchanged (they extend the per-platform bases
transitively). Net -290 lines.
